Fixes issue #2764 - Move all Server integration tests into /test/server subdirectory
