+++ test/pom.xml
@@ -28,11 +28,7 @@
         <module>micro</module>
         <module>mojarra2</module>
         <module>server</module>
-        <module>server-programmatic</module>
-        <module>server-slim</module>
         <module>soteria</module>
-        <module>wasp</module>
-        <module>wasp2</module>
         <module>webxml</module>
         <module>weld</module>
         <module>webapp</module>
