--- test/server-programmatic/src/main/webapp/WEB-INF/web.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<web-app xmlns="https://jakarta.ee/xml/ns/jakartaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/web-app_5_0.xsd"
-         version="5.0">
-    <servlet>
-        <servlet-name>ServerProgrammaticServlet</servlet-name>
-        <servlet-class>cloud.piranha.test.server.programmatic.ServerProgrammaticServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>ServerProgrammaticServlet</servlet-name>
-        <url-pattern>/index.html</url-pattern>
-    </servlet-mapping>
-    <session-config>
-        <session-timeout>30</session-timeout>
-    </session-config>
-</web-app>
