--- test/server-slim/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,0 @@
-cloud.piranha.test.server.slim.ServerSlimInitializer
