+++ test/server/wasp/src/main/webapp/WEB-INF/tld/hellotld.tld
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<taglib version="2.1" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-jsptaglibrary_2_1.xsd">
+  <tlib-version>1.0</tlib-version>
+  <short-name>hellotld</short-name>
+  <uri>/WEB-INF/tld/hellotld</uri>
+  <tag>
+    <name>hellotld</name>
+    <tag-class>wasp.HelloTldTag</tag-class>
+    <body-content>JSP</body-content>
+  </tag>
+</taglib>
